finished CRUD requests and services for all entities

+++ src/main/java/com/example/proiect_db/Model/Admission.java
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 public class Admission {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "patient_id", nullable = false)
@@ -36,7 +36,7 @@ public class Admission {
     public Admission() {
     }
 
-    public Admission(Long id, Patient patient, Room room, LocalDateTime admissionDate, LocalDateTime dischargeDate, String reason, AdmissionStatus status) {
+    public Admission(Integer id, Patient patient, Room room, LocalDateTime admissionDate, LocalDateTime dischargeDate, String reason, AdmissionStatus status) {
         this.id = id;
         this.patient = patient;
         this.room = room;
@@ -46,11 +46,11 @@ public class Admission {
         this.status = status;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
